+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Editor
   - Support `workspace/willRenameFiles`, renaming namespaces and all its references when a file is renamed.
+  - Don't save cache when classpath lookup failed.
   
 This release was supported by [Clojurists Together](https://www.clojuriststogether.org/)